Rename mc-util-test-vector and simplify code
