+++ util/test-vector/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "mc-util-test-vectors"
+name = "mc-util-test-vector"
 version = "1.0.0"
 authors = ["MobileCoin"]
 edition = "2018"
