--- util/test-vectors/src/lib.rs
@@ -1,7 +1,0 @@
-mod test_vector;
-mod test_vector_reader;
-mod test_vector_writer;
-
-pub use test_vector::*;
-pub use test_vector_reader::*;
-pub use test_vector_writer::*;
