--- util/test-vectors/src/test_vector.rs
@@ -1,6 +1,0 @@
-pub trait TestVector: Sized {
-    const FILE_NAME: &'static str;
-    const MODULE_SUBDIR: &'static str;
-
-    fn generate() -> Vec<Self>;
-}
